don't display delete link for first entry of issue

+++ src/renderingHtml.cpp
@@ -1290,7 +1290,8 @@ void RHtml::printIssue(const ContextParameters &ctx, const Issue &issue)
 
         // delete button
         time_t delta = time(0) - ee.ctime;
-        if ( (delta < DELETE_DELAY_S) && (ee.author == ctx.username) && (e == issue.latest) ) {
+        if ( (delta < DELETE_DELAY_S) && (ee.author == ctx.username) &&
+             (e == issue.latest) && e->prev) {
             // entry was created less than 10 minutes ago, and by same user, and is latest in the issue
             mg_printf(conn, "<a href=\"#\" class=\"sm_entry_delete\" title=\"Delete this entry (at most %d minutes after posting)\" ", (DELETE_DELAY_S/60));
             mg_printf(conn, " onclick=\"deleteEntry('/%s/issues/%s', '%s');return false;\">\n",
